fix bad case
