+++ gluten-substrait/src/main/scala/org/apache/gluten/extension/columnar/rewrite/PullOutPreProject.scala
@@ -45,6 +45,14 @@ object PullOutPreProject extends RewriteSingleNode with PullOutProjectHelper {
       case take: TakeOrderedAndProjectExec =>
         take.sortOrder.exists(o => isNotAttribute(o.child))
       case agg: BaseAggregateExec =>
+        val pullOutHelper =
+          BackendsApiManager.getSparkPlanExecApiInstance.genHashAggregateExecPullOutHelper(
+            agg.aggregateExpressions,
+            agg.aggregateAttributes,
+            false)
+        val allAggregateResultAttributes =
+          pullOutHelper.allAggregateResultAttributes(agg.groupingExpressions)
+        agg.child.output.size != allAggregateResultAttributes.size ||
         agg.groupingExpressions.exists(isNotAttribute) ||
         agg.aggregateExpressions.exists {
           expr =>
