+++ gluten-ut/spark33/src/test/scala/org/apache/spark/sql/hive/execution/GlutenHiveSQLQueryCHSuite.scala
@@ -123,49 +123,4 @@ class GlutenHiveSQLQueryCHSuite extends GlutenHiveSQLQuerySuiteBase {
       purge = false)
   }
 
-  testGluten("GLUTEN-7096: Same names in group by may cause exception") {
-    sql("create table if not exists test_7096 (day string, rtime int, uid string, owner string)")
-    sql("insert into test_7096 values ('2024-09-01', 123, 'user1', 'owner1')")
-    sql("insert into test_7096 values ('2024-09-01', 123, 'user1', 'owner1')")
-    sql("insert into test_7096 values ('2024-09-02', 567, 'user2', 'owner2')")
-    val query =
-      """
-        |select days, rtime, uid, uid, day1
-        |from (
-        | select day1 as days, rtime, uid, uid, day1
-        | from (
-        |   select distinct coalesce(day, "today") as day1, rtime, uid, owner
-        |   from test_7096 where day = '2024-09-01'
-        | )) group by days, rtime, guest_uid, owner_uid, day1
-        |""".stripMargin
-    val df = sql(query)
-    checkAnswer(df, Seq(Row("2024-09-01", 123, "user1", "owner1", "2024-09-01")))
-    spark.sessionState.catalog.dropTable(
-      TableIdentifier("test_7096"),
-      ignoreIfNotExists = true,
-      purge = false)
-  }
-
-  testGluten("GLUTEN-7096: Same names with different qualifier in group by may cause exception") {
-    sql("create table if not exists test_7096 (day string, rtime int, uid string, owner string)")
-    sql("insert into test_7096 values ('2024-09-01', 123, 'user1', 'owner1')")
-    sql("insert into test_7096 values ('2024-09-01', 123, 'user1', 'owner1')")
-    sql("insert into test_7096 values ('2024-09-02', 567, 'user2', 'owner2')")
-    val query =
-      """
-        |select days, rtime, uid, uid, day1
-        |from (
-        | select day1 as days, rtime, uid, uid, day1
-        | from (
-        |   select distinct coalesce(day, "today") as day1, rtime, uid, owner
-        |   from test_7096 where day = '2024-09-01'
-        | ) t1 ) t2 group by days, rtime, guest_uid, owner_uid, day1
-        |""".stripMargin
-    val df = sql(query)
-    checkAnswer(df, Seq(Row("2024-09-01", 123, "user1", "owner1", "2024-09-01")))
-    spark.sessionState.catalog.dropTable(
-      TableIdentifier("test_7096"),
-      ignoreIfNotExists = true,
-      purge = false)
-  }
 }
